\name{barplot.getx}
\alias{barplot.getx}
\title{Function to tell you where the middle of each bar is in a barplot.}
\usage{
  barplot.getx(height, width = 1, space = NULL,
    beside = FALSE, horiz = FALSE, ...)
}
\arguments{
  \item{height}{see \code{\link{barplot}}}

  \item{width}{see \code{\link{barplot}}}

  \item{space}{see \code{\link{barplot}}}

  \item{beside}{see \code{\link{barplot}}}

  \item{horiz}{see \code{\link{barplot}}}

  \item{\dots}{see \code{\link{barplot}}}
}
\value{
  a vector of midpoints of each bar in a barplot
}
\description{
  Function to tell you where the middle of each bar is in a
  barplot.
}
\author{
  Mark Cowley, 8/1/07
}

