\name{colour2hex}
\alias{colour2hex}
\title{Convert "col" format to "rgb" format.}
\usage{
  colour2hex(x = NULL)
}
\arguments{
  \item{x}{a vector of colours in the colour format:
  \dQuote{white}}
}
\description{
  ie from "White" to "#FFFFFF"
}
\examples{
colour2hex("White")
# [1] "#FFFFFF"
}
\author{
  Mark Cowley, 29/3/07
}

