\name{colour2rgb}
\alias{colour2rgb}
\title{Convert "col" format to "rgb" format.}
\usage{
  colour2rgb(x = NULL)
}
\arguments{
  \item{x}{a vector of colours in the colour format:
  \dQuote{white}}
}
\description{
  ie from "White" to c(255,255,255)
}
\examples{
colour2rgb("White")
# [1] c(255,255,255)
}
\author{
  Mark Cowley, 29/3/07
}

