\name{colours.hex}
\alias{colours.hex}
\title{Extend \code{\link[grDevices]{colours}} to return the hex values of the named colours. see
\code{\link{colours}}}
\usage{
  colours.hex()
}
\value{
  a vector of hex values such as \dQuote{#FF0EF0}
}
\description{
  Extend \code{\link[grDevices]{colours}} to return the hex
  values of the named colours. see \code{\link{colours}}
}
\author{
  Mark Cowley, 29/3/07
}

