\name{colours.rgb}
\alias{colours.rgb}
\title{Extend \code{\link[grDevices]{colours}} to return the rgb values of the named colours. see
\code{\link{colours}}}
\usage{
  colours.rgb()
}
\value{
  a vector of hex values such as ?
}
\description{
  Extend \code{\link[grDevices]{colours}} to return the rgb
  values of the named colours. see \code{\link{colours}}
}
\author{
  Mark Cowley, 29/3/07
}

