\name{density.peak}
\alias{density.peak}
\alias{get.density.peaks}
\title{Find the element at the peak of the density distribution}
\usage{
  density.peak(data)

  get.density.peaks(data)
}
\arguments{
  \item{data}{a matrix or vector of data}
}
\description{
  Find the element at the peak of the density distribution
}
\author{
  Mark Cowley, 12 July 2005
}

