\name{density_peak}
\alias{density_peak}
\alias{get_density_peaks}
\title{Find the element at the peak of the density distribution}
\usage{
  density_peak(data)

  get_density_peaks(data)
}
\arguments{
  \item{data}{a matrix or vector of data}
}
\description{
  Find the element at the peak of the density distribution
}
\author{
  Mark Cowley, 12 July 2005
}

