\name{image.table}
\alias{image.table}
\title{image plot of a table}
\usage{
  image.table(x, main = "Heatmap", col = NULL, zlim = NULL,
    legend = TRUE, unlog.legend = FALSE,
    legend.pos = "bottom", legend.main = "", na.rm = 0,
    xlab = "column index", xlabels = NULL,
    xlabels.pos = "bottom", ylab = "row index",
    ylabels = NULL, ylabels.pos = "left",
    row.descriptions = NULL, grid.colour = "white", ...)
}
\arguments{
  \item{x}{a \code{matrix} (or \code{data.frame}) of
  numbers}

  \item{legend}{add a color legend to RHS of plot}

  \item{col}{a list of colors such as that generated by
  \code{\link{rainbow}}, \code{\link{heat.colors}},
  \code{\link{topo.colors}}, \code{\link{terrain.colors}}
  or similar functions.}

  \item{zlim}{the minimum and maximum \code{z} values for
  which colors should be plotted, defaulting to the range
  of the finite values of \code{z}. Each of the given
  colors will be used to color an equispaced interval of
  this range. The \emph{midpoints} of the intervals cover
  the range, so that values just outside the range will be
  plotted.}
}
\value{
  an image with x[1,1] in top left, and x[nrow(x), ncol(x)]
  in bottom left corner Todo: get legend=TRUE to work when
  layout or mfrow/mfcol has prev. been set.
}
\description{
  Function to plot a \code{matrix} or \code{data.frame} as
  an image such that cell [1,1] is in top left of the
  image, and cell \code{[nrow, ncol]} is in bottom right
  corner, ie in the same orientation as \code{print} or
  \code{write.table} would show it. Note that
  \code{\link{image}} rotates the image along the diagonal.
}
\author{
  Mark Cowley, 2 June 2005
}

