\name{plot}
\alias{plot}
\alias{plot,ExpressionSet,ExpressionSet-method}
\title{xy-plot of data from 2 ExpressionSet's
Note only the first feature from each ExpressionSet are plotted.}
\description{
  xy-plot of data from 2 ExpressionSet's Note only the
  first feature from each ExpressionSet are plotted.
}

