\name{plot.pch}
\alias{plot.pch}
\title{plot the print characters}
\usage{
  plot.pch(file.name = NULL)
}
\arguments{
  \item{file.name}{the path to a pdf destination file, or
  \code{NULL}}
}
\value{
  none. makes a plot
}
\description{
  plot the print characters
}
\author{
  Mark Cowley, 2012-07-16
}

