\name{safe.brewer.pal}
\alias{safe.brewer.pal}
\title{Allows use of brewer.pal with less than 3 colours}
\usage{
  safe.brewer.pal(n, name)
}
\arguments{
  \item{n}{Number of different colors in the palette,
  minimum 1, maximum depending on palette}

  \item{name}{A palette name from the lists at ?brewer.pal}
}
\value{
  a vector of N colours
}
\description{
  Allows use of brewer.pal with less than 3 colours
}
\author{
  Mark Cowley, 2010-09-02
}

