\name{wrap.title}
\alias{wrap.title}
\title{Function to wrap long titles (from plots) by splitting at 'limit' characters
and inserting new lines}
\usage{
  wrap.title(text, limit = 28, maxrow = 3)
}
\description{
  Function to wrap long titles (from plots) by splitting at
  'limit' characters and inserting new lines
}
\author{
  Mark Cowley, 2009-03-23
}

